Depot final tentative 2

+++ eStage/src/components/OffersDashboard.vue
@@ -23,7 +23,7 @@
       </div>
 
       <!-- Grid header -->
-      <div className="hidden md:grid mt-20 grid-cols-4 gap-4">
+      <div className="hidden md:grid mt-20 grid-cols-4 gap-0">
         <div>Profil</div>
         <div>Ville</div>
         <div>Date</div>
@@ -31,24 +31,26 @@
       </div>
 
       <!-- Grid -->
-      <div className="mt-6 md:grid md:grid-cols-4 gap-4" v-for="offre in offres" :key="offre._id">
-        <div className="flex">
-          <div class="bg-red-600 w-1 h-12 mr-2 md:mr-6"></div>
-          <div className="hidden md:block bg-red-600 p-3 w-12 rounded-lg">
+      <div className="mt-6 md:grid md:grid-cols-4 gap-0" v-for="offre in offres" :key="offre._id">
+        <div className="flex md:border-b-4 md:p-1 md:border-slate-300">
+          <div className="bg-red-600 w-1 h-12 mr-2 md:mr-6"></div>
+          <div className="hidden md:block bg-red-600 p-3 w-12 h-12 rounded-lg">
             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
               <path fill="#ffffff"
                 d="M32 32H480c17.7 0 32 14.3 32 32V96c0 17.7-14.3 32-32 32H32C14.3 128 0 113.7 0 96V64C0 46.3 14.3 32 32 32zm0 128H480V416c0 35.3-28.7 64-64 64H96c-35.3 0-64-28.7-64-64V160zm128 80c0 8.8 7.2 16 16 16H336c8.8 0 16-7.2 16-16s-7.2-16-16-16H176c-8.8 0-16 7.2-16 16z" />
             </svg>
           </div>
-          <div className="text-black text-xs md:ml-2 my-auto">
-            <h3 className="font-semibold">{{ offre.title }}</h3>
-            <h4>{{ offre.description }}</h4>
+          <div className="text-black md:ml-2 my-auto">
+            <h3 className="font-semibold  text-xl">{{ offre.title }}</h3>
+            <h4 className="text-lg">{{ offre.description }}</h4>
           </div>
         </div>
-        <div className="text-black text-xs flex items-center">{{ offre.enterprise.city }}</div>
-        <div className="text-black text-xs flex items-center">{{ new Date(offre.startDate).toLocaleDateString('fr-CA')
+        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{
+        offre.enterprise.city }}</div>
+        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{ new
+        Date(offre.startDate).toLocaleDateString('fr-CA')
           }}</div>
-        <div className="text-xs flex justify-between items-center">
+        <div className="text-lg flex justify-between items-center md:border-b-4 md:p-1 md:border-slate-300">
           <button @click="accepterOffre(offre)" className="w-6 md:w-8 lg:hidden xl:hidden">
             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512">
               <path fill="#059650"
@@ -88,7 +90,6 @@
             </router-link>
           </button>
         </div>
-        <hr className="w-full bg-slate-300 h-1">
       </div>
     </div>
   </section>
@@ -117,12 +118,14 @@ const accepterOffre = async (offre) => {
 };
 
 const supprimerOffre = async (id) => {
-  // Appelle la méthode Delete du service en passant l'identifiant de l'offre
-  try {
-    const response = await service.Delete(id);
-    console.log('Offre supprimée avec succès :', response);
-  } catch (error) {
-    console.error(`Erreur lors de la suppression de l'offre :`, error);
+  if (confirm("Desirez-vous vraiment supprimer cela?")) {
+    // Appelle la méthode Delete du service en passant l'identifiant de l'offre
+    try {
+      const response = await service.Delete(id);
+      console.log('Offre supprimée avec succès :', response);
+    } catch (error) {
+      console.error(`Erreur lors de la suppression de l'offre :`, error);
+    }
   }
 };
